refactor(cli): extract PlaywrightProjectBundler with a shared cache
Every PlaywrightCheck called bundlePlayWrightProject(config, include),
which loads the Playwright config module, resolves the Playwright version,
and globs the project directory. Projects that generate many checks from
one shared codebase call it with identical arguments, repeating all of
that work per check.

Move bundlePlayWrightProject and its Playwright-only helpers
(getPlaywrightVersionFromPackage, getAutoIncludes, normalizeVersion) out
of services/util.ts into a new PlaywrightProjectBundler class that
memoizes the result per (config, include). The instance is shared via
Session.getPlaywrightProjectBundler() and cleared in Session.reset(),
mirroring Session.getPlaywrightParser(). The cache stores the in-flight
promise, so concurrent callers also share one computation. findFilesWithPattern
stays in util.ts (used elsewhere). Helper unit tests move alongside the
code into playwright-project-bundler.spec.ts.
